re-dumped the ROMs
I was able to re-dump the ROMs with a rebuilt ROM reader, and there are
several places that came out cleaner and make more sense now. The dumps
were also consistent, so I suspect any weirdness that’s left in the
ROMs is representative of what’s actually on the chips.